requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-CORS==5.0.1
Flask-Mail==0.10.0
PyMySQL==1.1.1
PyJWT==2.10.1
bcrypt==4.2.1
python-dotenv==1.1.0
pandas
openpyxl==3.1.5
cryptography==44.0.2
